[ntt] Move reference implementation to its own module
stack-info: PR: https://github.com/IrreducibleOSS/binius64/pull/1292, branch: stack/main/jimpo/jimpo/ntt-refactors/1
